Change default dataloader timeout to be the max of its sources
Without this, it's easy for people to configure a large timeout on a
source and not realise that the supervising process has killed it
prematurely by exiting on a default 15_000 timeout.

Note that this throws away the concept of setting a timeout on the
`Dataloader` level in. I initially implemented this allowing the timeout
to be set, but this ultimately makes no sense. The only consequence of
allowing people to set a lower timeout on the `Dataloader` than exists
on their `Source` is to open ourselves up to a whole new class of bugs,
there is no benefit to doing this. So just don't allow it :).