Add timeout metric for Dataloader.Source protocol
This may break existing protocol implementations. However this is
probably desired as the alternative would be to set some sort of
arbitrary default.

An arbitrary default may be dangerous in case anyone's implementing the
protocol and making assumptions that timeouts are handled (for example)
by `Dataloader.options`. If we set a default, they may not realise that
their timeouts have changed from what they were previously expecting.

Ultimately, we now rely directly on timeouts set on the source in order
to run the `Dataloader`, so making a breaking change highlights the
importance of this.