Add error handling when sources fail to load data
Previously, if any errors occurred when a source tried to load data,
this would get logged and a `nil` would get returned.

The danger of this is that this means errors may go unnoticed in
production environments (e.g. timeouts to third parties) as a `nil`
result may be a reasonable one depending on the `Source` and data.

This provides two additional configurable methods of error handling:

* `:raise_on_error` - this will raise a `Datlaoader.GetError` when one
  of the `get` methods are called. This is now the default behaviour
  for handling errors
* `:tuples` - this changes the `get/4`/`get_many/4` methods to return
  `:ok`/`:error` tuples instead of just the value. This frees up the
  caller to handle errors any way they see fit