Fixes issue #74
Handle case where Dataloader.KV has some results and is asked
to load an id that is neither present in 'results' or 'batches'.

In this case Map.get/2 would return `nil`, which was falling
through and being treated as if the id was found in the results.

This replaces Map.get/2 with Map.get/3 so that we can provide
an explicit value to pattern-match on (:error).