Refactor Map.get/3 into Map.fetch/2
Calling Map.get/3 with :error as the third argument is
almost the same as calling Map.fetch/2 except that non-errors
are wrapped in :ok tuples.

This more clearly delinates the cases where a result is found
and the case where a result is not found.  It also shows that
{:error, reason} is expected to be mapped to 'id' in some
situations.