Load rows for limit/offset queries with lateral join
Queries that do not rely on limit/offset continue to use
the existing approach, where a simple where clause is added
to the query.

Complex queries are converted into a lateral subquery.
The outer query contains the inputs.
Ideally the query would use `unnest` or a similar function
to efficiently load an array of values into a query, but
Ecto doesn't offer such functionality in a portable way.

The workaround here is to run a distinct query that selects
the column from the queryable where the value is any of the
inputs.