Fix duplicate records from lateral preload
This can happen when there is a through association like
A -has many-> B -belongs to-> C

When loading the A -> C association using a join, duplicate entries for C
were being returned.

The fix is to throw a distinct() on the inner query.
There may be a more elegant approach to detecting when
the distinct is necessary, but at least the results are correct now.