Fix for custom `distinct` handling in laterals
Ecto represents distinct clauses in a query as a
`%Ecto.Query.QueryExpr{}` struct. In Elixir, guards pass only when the
evaluated condition is true but since the distinct struct is not explicitly true, the guard fails to pass.

This update reverses the guard logic so that it passes as long as the distinct value is not nil.